% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvfSeq.R
\name{convert_complex_exp}
\alias{convert_complex_exp}
\title{convert_complex_exp: Convert a complex number from polar to exponential form}
\usage{
convert_complex_exp(z)
}
\arguments{
\item{z}{A complex type object as defined by R.}
}
\value{
Returns a vector containing the magnitude of z and the angle in radians corresponding to its expenential form.
}
\description{
Converts a complex number from polar to exponential form.
}
\examples{
convert_complex_exp(2-2i)
convert_complex_exp(-3+4i)
}
\author{
Aaron Khoo
}
