% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvfSeq.R
\name{spec_density}
\alias{spec_density}
\title{spec_density: Spectral density values}
\usage{
spec_density(s, lowerBound = -0.5, upperBound = 0.5, step = 0.1)
}
\arguments{
\item{s}{numeric vector, first row of toeplitz matrix.}

\item{lowerBound}{double, lower bound of frequency range.}

\item{upperBound}{double, upper bound of frequency range.}

\item{step}{double, increment for loop.}
}
\value{
Returns a numeric vector containing the values of the spectral distribution at each frequency index.
}
\description{
Calculates the spectral density of a sequence for a given range of frequencies assuming it is the autocovariance function (acvf) of a weakly stationary process.
}
\examples{
spec_density(s=c(1.25, 0.5), lowerBound=-0.5, upperBound=0.5)
spec_density(s=c(2, 1, 0.5, -0.2), lowerBound=-0.5, upperBound=0.5)
}
\author{
Aaron Khoo
}
