% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvfSeq.R
\name{spec_density_plot}
\alias{spec_density_plot}
\title{spec_density_plot: Spectral density plot}
\usage{
spec_density_plot(s, lowerBound = -0.5, upperBound = 0.5, step = 0.1)
}
\arguments{
\item{s}{numeric vector, first row of toeplitz matrix.}

\item{lowerBound}{double, lower bound of frequency range.}

\item{upperBound}{double, upper bound of frequency range.}

\item{step}{double, increment for loop.}
}
\value{
Returns the spectral density plot.
}
\description{
Plots values of the spectral density of a sequence for a given frequency range.
}
\examples{
spec_density_plot(s=c(1.25, 0.5), lowerBound=-0.5, upperBound=0.5)
spec_density_plot(s=c(1, 0, 0.5, -0.2), lowerBound=-0.5, upperBound=0.5)
}
\author{
Aaron Khoo
}
