% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvfSeq.R
\name{spec_density_w}
\alias{spec_density_w}
\title{spec_density_w: Helper function}
\usage{
spec_density_w(s, w)
}
\arguments{
\item{s}{numeric sequence, first row of toeplitz matrix.}

\item{w}{double, frequency.}
}
\value{
Returns value of spectral density at the given frequency.
}
\description{
Calculates the spectral density value for a given frequency.
}
\author{
Aaron Khoo
}
