% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acvfSeq.R
\name{tmatrix_psd}
\alias{tmatrix_psd}
\title{tmatrix_psd: Checks the positive semidefinite property of a sequence}
\usage{
tmatrix_psd(n = 100, s)
}
\arguments{
\item{n}{numeric, the maximum dimensions of the toeplitz matrix to check.}

\item{s}{numeric vector, the sequence which forms the first row of the toeplitz matrix.}
}
\value{
Returns the dimensions of the matrix at which the sequence is found to be not positive semidefinite. If it cannot determine for the given n, returns -1.
}
\description{
Generates a toeplitz matrix with the given sequence as its first row and checks it it is not positive semidefinite. Sequences must be at least of length 2.
}
\details{
After checking the initial matrix generated by s, the function creates a new sequence by adding 0 to the existing one. It repeats till the dimensions of the matrix generated is n by n.
}
\examples{
tmatrix_psd(100, c(1, 0, -0.4, -0.2))
tmatrix_psd(2, c(6, 3))
}
\author{
Aaron Khoo
}
